+++ rappi4/src/Pages/Cadastro.js
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom";
 import axios from 'axios';
 import  useForm  from "../Hooks/UseForm"
 import { BASE_URL } from "../Constants/urls"
+import { vaiParaLogin } from "../Router/RouteFunctions";
 
 
 const Cadastro = () => {
@@ -17,7 +18,7 @@ const Cadastro = () => {
   const onSubmitForm = (event) => {
     event.preventDefault()
     console.log(form)
-    //signUp(form, clear, navigate)
+    clear()
   }
 
   const onSubmitCadastro = () => {
@@ -26,30 +27,30 @@ const Cadastro = () => {
     .then((res) => {
       console.log(res.data.token);
       localStorage.setItem("token", res.data.token);
-      navigate("/");
+      navigate("/endereco");
     })
     .catch((err) => {
       console.log(err.response.data);
-      alert("Usuário não cadastrado")
     });
   }
 
   return (
     <div>
       <h1>Cadastrar</h1>
+      <button onClick={() => vaiParaLogin(navigate)}>Voltar</button>
       <form onSubmit={onSubmitForm}>
         <input
           value={form.name}
           name={'name'}
           onChange={onChange}
-          placeholder="nome"
+          placeholder="Nome"
           required
           />
           <input
           value={form.email}
           name={'email'}
           onChange={onChange}
-          placeholder="email"
+          placeholder="E-mail"
           required
           type={"email"}
           />
@@ -59,14 +60,17 @@ const Cadastro = () => {
           onChange={onChange}
           placeholder="CPF"
           required
+          pattern={"\d{3}\.\d{3}\.\d{3}-\d{2}"}
+			    title={"Digite um CPF no formato: xxx.xxx.xxx-xx"}
           type={"cpf"}
           />
           <input
           value={form.password}
           name={'password'}
           onChange={onChange}
-          placeholder="senha"
+          placeholder="Senha"
           required
+          type={"password"}
           />
         
         <button
